Moved FluxEstimate to estimators.core.py
